requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.13.0
python-dotenv==1.0.1
pydantic==2.9.0
pydantic-settings==2.6.0
feedparser==6.0.11
httpx==0.25.2
beautifulsoup4==4.12.2
lxml==5.1.0
jinja2==3.1.2
pytest==7.4.3
pytest-asyncio==0.21.1

